edit

+++ Presentation/SE.Website/Models/Table1Model.cs
@@ -7,21 +7,21 @@ namespace Website.Models
 {
     public class Table1Model
     {
-        public decimal? A1 { get; set; }
-        public decimal? B1 { get; set; }
-        public decimal? C1 { get; set; }
-        public decimal? D1 { get; set; }
-        public decimal? E1 { get; set; }
-        public decimal? F1 { get; set; }
-        public decimal? G1 { get; set; }
-        public decimal? H1 { get; set; }
+        public double? A1 { get; set; }
+        public double? B1 { get; set; }
+        public double? C1 { get; set; }
+        public double? D1 { get; set; }
+        public double? E1 { get; set; }
+        public double? F1 { get; set; }
+        public double? G1 { get; set; }
+        public double? H1 { get; set; }
 
-        public decimal? A2 { get; set; }
-        public decimal? B2 { get; set; }
-        public decimal? C2 { get; set; }
-        public decimal? D2 { get; set; }
-        public decimal? E2 { get; set; }
-        public decimal? F2 { get; set; }
-        public decimal? G2 { get; set; }
+        public double? A2 { get; set; }
+        public double? B2 { get; set; }
+        public double? C2 { get; set; }
+        public double? D2 { get; set; }
+        public double? E2 { get; set; }
+        public double? F2 { get; set; }
+        public double? G2 { get; set; }
     }
 }
